Checkpoint 2 submission file.
This version is the final merger between the two files before the checkpoint. At this point, 3 levels are available with the option to restart each level and a progress bar. There are instructions on the main page. The character dies if he falls off the map or hits an object. The character is unable to travel infinity above the screen. The code has been cleaned up with PEP compliance.